New piet language extension and its interpreter
This language contains 6 new colors :
- dark orange, orange and light orange
- dark gray, gray, light gray

and 6 new operations :
- jump which pops the two first elements of the stack and jump the cursor's interpretor to the codel
which column is top value and line is second top value (line 1 is bottom line).
- huechange which pops the three first elements of the stack and change the hue of the codel
which column is top value and line is second top value. the number of steps along the hue cycle
between original color and final color of the codel is designated by the third top value. this
value can be non positive.
- lightchange which pops the three first elements of the stack and change the lightness of the codel
which column is top value and line is second top value. the number of steps along the lightness cycle
between original color and final color of the codel is designated by the third top value. this
value can be non positive.
- 2white which pops the two first elements of the stack, and change to white the color of the codel
which column is top value and line is second top value.
- 2black which pops the two first elements of the stack, and change to black the color of the codel
which column is top value and line is second top value.
- 2color which pops the four first elements of the stack, and change the color of the codel
which column is top value and line is second top value. The new color is a color which hue is
the third top value, and lightness is the fourth.